move etw_data to exporters/etw, wrap modules with if _WIN32
