+++ ci/do_ci.sh
@@ -112,30 +112,30 @@ EOF
   examples/plugin/load/load_plugin_example ${PLUGIN_DIR}/libexample_plugin.so /dev/null
   exit 0
 elif [[ "$1" == "bazel.test" ]]; then
-  bazel build $BAZEL_OPTIONS -- //... -//exporters/etw/...
-  bazel test $BAZEL_TEST_OPTIONS -- //... -//exporters/etw/...
+  bazel build $BAZEL_OPTIONS //...
+  bazel test $BAZEL_TEST_OPTIONS //...
   exit 0
 elif [[ "$1" == "bazel.legacy.test" ]]; then
   # we uses C++ future and async() function to test the Prometheus Exporter functionality,
   # that make this test always fail. ignore Prometheus exporter here.
-  bazel build $BAZEL_OPTIONS -- //... -//exporters/otlp/... -//exporters/prometheus/... -//exporters/etw/...
-  bazel test $BAZEL_TEST_OPTIONS -- //... -//exporters/otlp/... -//exporters/prometheus/... -//exporters/etw/...
+  bazel build $BAZEL_OPTIONS -- //... -//exporters/otlp/... -//exporters/prometheus/...
+  bazel test $BAZEL_TEST_OPTIONS -- //... -//exporters/otlp/... -//exporters/prometheus/...
   exit 0
 elif [[ "$1" == "bazel.noexcept" ]]; then
   # there are some exceptions and error handling code from the Prometheus Client
   # that make this test always fail. ignore Prometheus exporter in the noexcept here.
-  bazel build --copt=-fno-exceptions $BAZEL_OPTIONS -- //... -//exporters/prometheus/... -//exporters/etw/...
-  bazel test --copt=-fno-exceptions $BAZEL_TEST_OPTIONS -- //... -//exporters/prometheus/... -//exporters/etw/...
+  bazel build --copt=-fno-exceptions $BAZEL_OPTIONS -- //... -//exporters/prometheus/...
+  bazel test --copt=-fno-exceptions $BAZEL_TEST_OPTIONS -- //... -//exporters/prometheus/...
   exit 0
 elif [[ "$1" == "bazel.asan" ]]; then
-  bazel test --config=asan $BAZEL_TEST_OPTIONS -- //... -//exporters/etw/...
+  bazel test --config=asan $BAZEL_TEST_OPTIONS //...
   exit 0
 elif [[ "$1" == "bazel.tsan" ]]; then
-  bazel test --config=tsan $BAZEL_TEST_OPTIONS -- //... -//exporters/etw/...
+  bazel test --config=tsan $BAZEL_TEST_OPTIONS //...
   exit 0
 elif [[ "$1" == "bazel.valgrind" ]]; then
-  bazel build $BAZEL_OPTIONS -- //... -//exporters/etw/...
-  bazel test --run_under="/usr/bin/valgrind --leak-check=full --error-exitcode=1 --suppressions=\"${SRC_DIR}/ci/valgrind-suppressions\"" $BAZEL_TEST_OPTIONS -- //... -//exporters/etw/...
+  bazel build $BAZEL_OPTIONS //...
+  bazel test --run_under="/usr/bin/valgrind --leak-check=full --error-exitcode=1 --suppressions=\"${SRC_DIR}/ci/valgrind-suppressions\"" $BAZEL_TEST_OPTIONS //...
   exit 0
 elif [[ "$1" == "benchmark" ]]; then
   [ -z "${BENCHMARK_DIR}" ] && export BENCHMARK_DIR=$HOME/benchmark
